% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{d_node}
\alias{d_node}
\title{Create a decision node.}
\usage{
d_node(children, name = "Decision")
}
\arguments{
\item{children}{List. A list of nodes that are the
alternatives of this decision.}

\item{name}{Character. The name of this outcome}
}
\value{
A node (\code{list}) that has a \code{value}, \code{name}, and the
  selected \code{decision}.
}
\description{
\code{d_node} returns a decision node that takes the max value of it's
children as its value. A decision node can have multiple children.
}
\examples{
# Generate a list of some outcome nodes to be children
children <- list(
  o_node(v = -10, name = "Rain"),
  o_node(v = 10, name = "Sunny")
)
d <- d_node(children = children, name = "Weather")
str(u)
}
\seealso{
Other node generating functions: \code{\link{o_node}},
  \code{\link{u_node}}
}
