% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datree.R
\docType{package}
\name{datree}
\alias{datree}
\alias{datree-package}
\title{datree: A package for doing decision analysis in R.}
\description{
The \code{datree} package provides functions for building, computing, and
analyzing decision analysis trees in R.
}
