% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{decide}
\alias{decide}
\title{Generate data frame of alternative values for a decision node}
\usage{
decide(node, data = NULL)
}
\arguments{
\item{node}{\code{d_node}. The decision node to evaluate.}

\item{data}{List-like. The data to be used when evaluating the value
functions. Depending on the value functions, omitting the data may result
in an error.}
}
\value{
Data frame of alternatives and respective values.
}
\description{
\code{decide} uses \code{lazyeval} to evaluate the value functions of a
decision node's children, i.e., alternatives.
}
\examples{
a <- o_node(v = ~ .data$x + y, "Add")
b <- o_node(v = ~ .data$x * y, "Multiply")
d <- d_node(children = list(a, b))
y <- 10
decide(d, data = list(x = 10))
decide(d, data = list(x = 1))
}
\seealso{
Other node processing functions: \code{\link{evaluate}}
}
