% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate the value of a node}
\usage{
evaluate(node, data = NULL, verbose = FALSE)
}
\arguments{
\item{node}{Node list. The node to evaluate.}

\item{data}{List-like. The data to be used when evaluating the value
function. Depending on the value function, omitting the data may result in
an error.}

\item{verbose}{Logical. Whether or not to print the full value formula.}
}
\value{
Evaluated numeric value.
}
\description{
\code{evaluate} uses \code{lazyeval} to evaluate the value function of any
given node. Essentially a shortcut for calling \code{lazyeval::f_eval} on the
\code{$value} object of a node.
}
\examples{
v <- o_node(v = ~ .data$x + y)
y <- 10
evaluate(v, data = list(x = 10))
evaluate(v, data = list(x = 1))
}
\seealso{
Other node processing functions: \code{\link{decide}}
}
