% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{o_node}
\alias{o_node}
\title{Create an outcome (value) node.}
\usage{
o_node(v, name = "Outcome")
}
\arguments{
\item{v}{Numeric. The value assigned to this outcome.}

\item{name}{Character. The name of this outcome}
}
\value{
A node (\code{list}) that has a \code{value} and \code{name}.
}
\description{
\code{o_node} returns a value node that usually acts as a outcome leaf.
A value node should only be associated with a single outcome value.
}
\examples{
v <- o_node(v = 5)
str(v)
}
\seealso{
Other node generating functions: \code{\link{d_node}},
  \code{\link{u_node}}
}
